assemblers/qemu: add support for format options and compression
Previously, format options and compression was hardcoded per format type.
This fact made experimenting with different settings hard. Additionally,
osbuild-composer is looking into switching the image for AWS to
streamOptimized VMDK, which is impossible to create with the current
osbuild capabilities.

This commit adds a new schema for specifying the image format. It allows users
to specify arbitrary options and whether to use compression. Backwards
compatibility with the old schema is still maintained.