sources: Add support for file contents to sources
This adds org.osbuild.blobs which stores the contents of files into the
manifest as ASCII encoded strings. It supports base16, base32, base64,
and base85 encoding. The files are indexed by the checksum of their
contents.

eg.
    "sources": {
        "org.osbuild.blobs": {
            "sha256:5ee3272ee76d8289126938c94411033af9f64f1b565e29c46907813252ba0460": "base64:W2Nvbm5lY3Rpb25dCmNvbm5lY3Rpb24uYXV0b2Nvbm5lY3Qtc2xhdmVzPTEK"
        }
    }

It is meant to be used for short text files, like configuration files,
not for large multi-megabytes files.