test: Move stages/copy test to stages/copy.files
This test tests org.osbuild.files as the source of the copy, rename it
to make it more obvious what it does.

This also simplifies the manifest to speed up the tests. There is no
need to do a full rpm install of the tree to test the copy stage.