test/data/manifests: use snapshots
Use snapshots from rpmrepo.osbuild.org, rather than a real repository.

In tests we want repositories to be persistent and immutable, which are
note guarantees real repositories usually give.

We only test with fedora, and only the released repositories, but with
the snapshots this could now be extended in a follow-up to use also updates and
modules for
completeness.