Makefile: mark '$(TEST_MANIFESTS_GEN)' target as .PHONY
Mark '$(TEST_MANIFESTS_GEN)' target as .PHONY.

Currently the `test-data` make target does not always trigger
regenerating of manifests used for testing various osbuild parts,
although it is marked as .PHONY. The reason that its dependency
'$(TEST_MANIFESTS_GEN)' is not marked as .PHONY and therefore it is not
run if the files already exist in repository.

Due to the above reason, CI is actually running `make --always-make
test-data` to always regenerate manifests.

Signed-off-by: Tomas Hozza <thozza@redhat.com>